Separate contexts for models; cut subclass tests.
With the lets outside the context, they weren't being reassigned to take
both model_name values. Move them into the context so we test both.

Testing that subclassing MultiparameterDateTime preserves parent
behavior is testing that inheritance works in our framework.
This is out of scope for our own unit tests.